Move the EmitC reference impl to its own folder
This moves all `include/emitc/emitc_*.h` header files to
`reference-implementation/include/emitc/`. Furthermore, the `emitc_`
prefix on the header files is removed.
