+++ CMakeLists.txt
@@ -106,3 +106,4 @@ add_subdirectory(include/emitc)
 add_subdirectory(lib)
 add_subdirectory(test)
 add_subdirectory(tools)
+add_subdirectory(reference-implementation)
