+++ reference-implementation/include/emitc/utility.h
@@ -12,8 +12,8 @@
 
 // This file defines the utility functions used in EmitC.
 
-#ifndef EMITC_EMITC_UTILITY_H
-#define EMITC_EMITC_UTILITY_H
+#ifndef EMITC_UTILITY_H
+#define EMITC_UTILITY_H
 
 #include <array>
 #include <cassert>
@@ -93,4 +93,4 @@ constexpr std::array<size_t, sizeof...(Shape)> unravel_index(size_t index) {
 } // namespace utility
 } // namespace emitc
 
-#endif // EMITC_EMITC_UTILITY_H
+#endif // EMITC_UTILITY_H
