+++ scripts/e2e_test.sh
@@ -14,7 +14,7 @@
 set -e
 
 if [[ $# -ne 7 ]] ; then
-  echo "Usage: $0 <path/to/model> <path/to/emitc/include/files> <path/to/emitc-opt> <compiler> <batch-size> <seed> <output_dir>"
+  echo "Usage: $0 <path/to/model> <path/to/emitc/reference-implementation/include/> <path/to/emitc-opt> <compiler> <batch-size> <seed> <output_dir>"
   echo
   echo "Both a keras and a tensorflow saved model is supported."
   echo
