+++ test/Conversion/arith-to-emitc.mlir
@@ -3,7 +3,7 @@
 // RUN: emitc-opt -insert-emitc-arith-include -convert-arith-to-emitc %s | FileCheck %s --check-prefixes=CHECK,CHECK-INCLUDE
 // RUN: emitc-opt -arith-to-emitc-pipeline %s | FileCheck %s --check-prefixes=CHECK,CHECK-INCLUDE
 
-// CHECK-INCLUDE: emitc.include "emitc_arith.h"
+// CHECK-INCLUDE: emitc.include "emitc/arith.h"
 
 func @arith_index_cast(%arg0: tensor<index>, %arg1: tensor<2xi32>, %arg2: tensor<2x2xi32>) -> tensor<2xindex> {
   %0 = "arith.index_cast"(%arg0) : ( tensor<index>) -> tensor<i32>
