+++ test/Conversion/mhlo-to-emitc.mlir
@@ -2,7 +2,7 @@
 // RUN: emitc-opt --insert-emitc-mhlo-include -convert-mhlo-region-ops-to-emitc -convert-mhlo-to-emitc %s | FileCheck %s  --check-prefixes=CHECK,CHECK-INCLUDE
 // RUN: emitc-opt -mhlo-to-emitc-pipeline %s | FileCheck %s --check-prefixes=CHECK,CHECK-INCLUDE
 
-// CHECK-INCLUDE: emitc.include "emitc_mhlo.h"
+// CHECK-INCLUDE: emitc.include "emitc/mhlo.h"
 
 // Nullary ops
