+++ test/Conversion/tensor-to-emitc.mlir
@@ -3,7 +3,7 @@
 // RUN: emitc-opt -insert-emitc-tensor-include -convert-tensor-to-emitc %s | FileCheck %s --check-prefixes=CHECK,CHECK-INCLUDE
 // RUN: emitc-opt -tensor-to-emitc-pipeline %s | FileCheck %s --check-prefixes=CHECK,CHECK-INCLUDE
 
-// CHECK-INCLUDE: emitc.include "emitc_tensor.h"
+// CHECK-INCLUDE: emitc.include "emitc/tensor.h"
 
 func @std_extract_element(%arg0: tensor<i32>, %arg1: tensor<2xi32> ) -> () {
   %0 = arith.constant 0 : index
