+++ test/Conversion/tosa-to-emitc.mlir
@@ -2,7 +2,7 @@
 // RUN: emitc-opt -insert-emitc-tosa-include -convert-tosa-to-emitc %s | FileCheck %s --check-prefixes=CHECK,CHECK-INCLUDE
 // RUN: emitc-opt -tosa-to-emitc-pipeline %s | FileCheck %s --check-prefixes=CHECK,CHECK-INCLUDE
 
-// CHECK-INCLUDE: emitc.include "emitc_tosa.h"
+// CHECK-INCLUDE: emitc.include "emitc/tosa.h"
 
 // Data node ops
